requirements.txt
pypdf
pydantic
langchain-openai
python-dotenv
langchain-groq
streamlit
universal-extractor-backend
pandas
